move: formatPhoneNumber uitl로 분리
