+++ src/pages/SignUpPage/PhoneNumberInput.tsx
@@ -1,6 +1,7 @@
 import { ChangeEvent, useCallback } from "react";
 import { TextField, Tooltip } from "@mui/material";
 import ErrorOutlineIcon from "@mui/icons-material/ErrorOutline";
+import { formatPhoneNumber } from "@utils/numberUtil";
 
 export interface PhoneNumberInputProps {
   phoneNumber: string;
@@ -16,43 +17,6 @@ const isValidPhoneNumber = (phone: string) => {
     /^01[0|1|6|7|8|9]-\d{3,4}-\d{4}$/.test(phone) &&
     (phone.length === 12 || phone.length === 13)
   );
-};
-
-const formatPhoneNumber = (input: string): string => {
-  const cleanedInput = input.replace(/[^0-9]/g, "");
-  let formattedNumber = "";
-
-  if (cleanedInput.startsWith("02")) {
-    if (cleanedInput.length <= 2) {
-      formattedNumber = cleanedInput;
-    } else if (cleanedInput.length <= 6) {
-      formattedNumber = cleanedInput.replace(/^(\d{2})(\d{0,4})/, "$1-$2");
-    } else {
-      formattedNumber = cleanedInput.replace(
-        /^(\d{2})(\d{4})(\d{4})/,
-        "$1-$2-$3"
-      );
-    }
-    if (formattedNumber.length > 11) {
-      formattedNumber = formattedNumber.slice(0, 11);
-    }
-  } else {
-    if (cleanedInput.length <= 3) {
-      formattedNumber = cleanedInput;
-    } else if (cleanedInput.length <= 7) {
-      formattedNumber = cleanedInput.replace(/^(\d{3})(\d{0,4})/, "$1-$2");
-    } else {
-      formattedNumber = cleanedInput.replace(
-        /^(\d{3})(\d{4})(\d{4})/,
-        "$1-$2-$3"
-      );
-    }
-    if (formattedNumber.length > 13) {
-      formattedNumber = formattedNumber.slice(0, 13);
-    }
-  }
-
-  return formattedNumber;
 };
 
 const PhoneNumberInput = ({
